Switched Datatypes in SoundFont & Player from Double to Float

+++ app/src/main/java/com/qfs/apres/soundfontplayer/PitchedBuffer.kt
@@ -4,23 +4,23 @@ import kotlin.math.abs
 import kotlin.math.max
 import kotlin.math.min
 
-class PitchedBuffer(val data: ShortArray, var pitch: Double) {
+class PitchedBuffer(val data: ShortArray, var pitch: Float) {
     val data_size = data.size
     val max = max(abs(data.min().toInt()), data.max().toInt())
-    var size = (data.size.toDouble() / this.pitch).toInt()
+    var size = (data.size.toFloat() / this.pitch).toInt()
 
     private var virtual_position: Int = 0
-    private val original_pitch: Double = this.pitch
+    private val original_pitch: Float = this.pitch
 
-    fun repitch(new_pitch: Double) {
+    fun repitch(new_pitch: Float) {
         this.pitch = this.original_pitch * new_pitch
-        this.size = (this.data_size.toDouble() / this.pitch).toInt()
+        this.size = (this.data_size.toFloat() / this.pitch).toInt()
         this.virtual_position = min(this.virtual_position, this.size - 1)
     }
 
     fun reset_pitch() {
         this.pitch = this.original_pitch
-        this.size = (this.data_size.toDouble() / this.pitch).toInt()
+        this.size = (this.data_size.toFloat() / this.pitch).toInt()
     }
 
     fun position(): Int {
@@ -32,7 +32,7 @@ class PitchedBuffer(val data: ShortArray, var pitch: Double) {
     }
 
     fun get(): Short {
-        return this.data[((this.virtual_position++).toDouble() * this.pitch).toInt()]
+        return this.data[((this.virtual_position++).toFloat() * this.pitch).toInt()]
     }
 
 }
